requirements.txt
Flask
flask-cors
python-dotenv
pymongo
werkzeug
PyJWT
